% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucdown.R
\docType{package}
\name{ucdown}
\alias{ucdown}
\alias{ucdown-package}
\title{ucdown: A package for creating undergraduate, Masters, and PhD theses
using R Markdown}
\description{
ucdown: A package for creating undergraduate, Masters, and PhD theses
using R Markdown
}
